Done:Edit for certificate and Access check for lessons and test

+++ KnowledgeVista/src/main/java/com/knowledgeVista/Course/Test/MuserTestActivity.java
@@ -40,7 +40,7 @@ public class MuserTestActivity {
 	    private CourseTest test;
 	    
 	    private LocalDate testDate;
-	    private Integer nthAttempt;
+	    private Long nthAttempt;
 	   
 	    private Double percentage;
